Improve handling of libs in wrapper compilers
Improve the logic of handling library (-l) references in the
wrapper compilers.  Previously, a duplicate in one of the LIB
strings would leave the left-most reference, which can lead to
missing symbols if a more right library also depends on the
duplicated library.  Instead, keep only the right-most reference.

Add a helper macro for list appending to avoid duplicating the
same 3 lines of shell code multiple times in opal_functions.m4

Signed-off-by: Brian Barrett <bbarrett@amazon.com>