Fix handling of configure component LDFLAGS/LIBS
Refactor the code to pull wrapper LDFLAGS and LIBS from component
configure code into the opal_mca system.  The previous code was a
bit too cute for my own good (read: overly complicated) and also
had a bug that it assumed $3 was not a shell variable (which is
not true if a component provides a configure script).  Properly
use AS_VAR_COPY to avoid that bug, and manually expand the LDFLAGS
and LIBS loops to simplify the code.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>
(cherry picked from commit badc3188cd05813b8006ac2cb3dae5b2832a328a)