Improve pkg-config wrapper compiler handling
Rewrite the pkg-config code to split LDFLAGS and LIBS detection
as well as to return the flags in an environment variable instead
of updating the wrapper compiler flags directly.  The split
keeps us from adding -L flags to LIBS (as we had been doing) and
the environment variable means that we can consider using
pkg-config in OPAL_CHECK_PACKAGE, since some of our dependencies
(PMIX) are getting more complicated over time and our package
checks will struggle to keep up.

Rewrite the wrapper compiler code in the hwloc, libevent, and
pmix checks to handle the new pkg-config interface.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>