Fix two bugs in OPAL_FLAGS_APPEND_MOVE
Fix two issues in OPAL_FLAGS_APPEND_MOVE, which impacted the static
library list in the wrapper compilers.  First, an assignment was
outside the conditional clause, resulting in the original array
being wiped out if the input did not contain any -l arguments.
Second, we did not have the default case to handle arguments that
did not begin with -I, -L, or -l.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>