HWLOC/Libevent: Handle non-pkg-config LIBS
Better handle setting the wrapper compiler LIBS for hwloc and
libevent when they are co-built and pkg-config is not available.
Previously, we would try to add the .la files to the wrapper compilers,
which is probably not what we wanted.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>
(cherry picked from commit 4a35214f1fedf70bf8e69a4f2088c2ea7ac3f6c3)