config/opal_config_asm.m4: trivial output label fix
"OMPI" was a misleading label in the configure output summary of the
atomics choice, and it was weird that there was a whole section just
for Atomics.  Instead, change the output to be "Atomics" in the
Miscellaneous section.  Specifically, the output changes from this:

```
Atomics
-----------------------
OMPI: BUILTIN_C11
```

to this:

```
Miscellaneous
-----------------------
...
Atomics: BUILTIN_C11
```

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>