ompi_fortran_get_alignment.m4: fix test
LOC() is not a universal Fortran function (perhaps it's a GNU
extension that at least some Fortran compilers have adopted?  It
doesn't seem to exist in the NAG compiler, at least).

Additionally, doing the pointer arithmetic is not technically valid
Fortran.  So make the test for mpi_f08 handle alignment be similar to
the other alignment tests: make instances of the variable in a common
block and use C to compute the alignment.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>