osc/rdma: Load all btls as alternate btls
See lengthy comment in the change, but this patch removes the
ability of users to specify a subset of available btls for use
by the osc rdma component.  The BTL interface was never designed
for such usage (which is why there is no similar option for the
OB1 PML) and it had clear places where it broke, so remove it.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>
(cherry picked from commit 61aca9e2de7d35c9016f1dc2cf1750c7ab2b559f)