btl/am-rdma: Disable underlying flush
Disable the underlying flush() call when using the am-rdma interfce.
The current code does not properly handle flushing between the
am-rdma and underlying code properly, so it is quite easy to end
up with flush not behaving the way users expect.  Since there
are currently no callers that rely on a btl having both flush and
am-rdma interfaces, explicitly disable the case for now.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>