ompi/attr: update attribute system to prepare for sessions
This commit updates how the attribute system is initialized. Instead of
being set up in MPI_Init it is initialized by the subsystems (communicator,
windows, datatypes) that use it. To keep the delta from the sessions PR
minimal the ompi_mpi_instance_retain() and ompi_mpi_instance_release()
functions are stubbed out as no-ops. This will be removed once the instance
system is added.

Signed-off-by: Nathan Hjelm <hjelmn@google.com>