osc/rdma: Simplify query logic
With the change to load all btls in the alternate case and the
removal of the logic to change priority based on alternate or
accelerated btls, there's no point in running the full btl
query in osc_rdma_component_query().  Instead, just verify
that there is some number of BTLs available, which is the
extent of testing in the alternate case.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>
(cherry picked from commit e93f041aa291d19c2c482023ae88ffc149c01395)