osc/rdma: Split btl storage based on mode
Split the storage of btls based on whether the window is using
accelerated or alternate btls.  This makes it more obvious when
the code has made assumptions about mode that may not be true
(such as the memory registration calls throughout the code that
assumed selected_btl[0] was the one true BTL).

Signed-off-by: Brian Barrett <bbarrett@amazon.com>