Change atomic inclusion to make C11 less unique
With the previous refactoring and breaking out the implementation
code into seperate headers, it is no longer required that the
atomic_stdc.h implementation be included "special" in the
atomics header.  Clean up the ordering of includes so that
every implementation follows the same include ordering.

The stdc implementations of the atomics is type independent
and we don't enforce inline wrappers because that results in
a massive number of warnings from opal_lifo due to its interesting
use of volatile.  This patch takes the do-no-harm approach, but
we need to clean up opal_lifo in a future patch.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>
(cherry picked from commit 057a6b5af61e38990a9a0fb34583b347b72fd8b1)