LL/SC: be flexible with ret type
Do not use a temporary variable for holding the loaded value
for ll/sc calls, eliminiating warnings about type mismatches
for places where the compiler can implicitly cast into a
register type anyway.  This likely removes a slight bit of
type safety, but also removes an ugly to fix warning about
the use of typeof() on recent LLVM compilers that would be
near impossible to fix and would hide all other warnings in
noise.  Given the relatively infrequent use of LL/SC and the
ability to pay close attention when using it anyway, this is
probably the more sane call.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>