Refactor atomic math interface
The atomic math interface has become required based on usage,
so embrace the requirement and remove all the gorp to make the
interface optional.

Split out the compatibility implementation into headers that can
be included as required by implementations, saving us from a
mess of #defines in atomic_impl.h.

Reorganize the implementation headers so that the code is in the same
order across implementations, for readability.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>