btl/am-rdma: Clean up warnings
Remove unused function warning for the unused size_t max function.

Remove unused variable length array argument in the header.  While
the existance of the variable length array argument itself isn't
a problem, that header is included in the middle of another structure,
which is a no-no.  Since we don't use the variable length array
member, simplest solution is to remove it.

Copy the same compare and swap type enforcement for the 8 byte CAS
case as other similar cases (both 4 byte CAS and other atomics).

Signed-off-by: Brian Barrett <bbarrett@amazon.com>