util: Fix write handling in multiple functions
Fix both a real (but unlikely) error and warning with the return code
of write().  Recent versions of Glibc tag the write() function to raise
a warning if the return value is ignored.  The issue is real, so do
a slightly better job of handling the issue and silence the warning.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>
(cherry picked from commit 6fea88974a6be1dfa974d3d4ca89ad8eac855554)