Make atomic cswap and swap mandatory
The compare and swap and swap atomic interfaces are de facto
mandatory, as no code actually checks for their existance before
using them (with the notable exception of the 128 bit cswap), so
make those interfaces mandatory and remove all the complex code
to detect implementations of them.

At the same time, reorder the implementation files to all have
consistent ordering of the cswap and swap implementations, to
ease maintenance and move to a implementation includes the
compatibility layer mode instead of the previous overly complex
all in one software implementation model.

Finally, remove the untyped macro wrappers as 1) no one uses
them and 2) they're a pain to maintain.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>