treematch: Fix warnings
Fix two classes of warnings in the treematch code.  First, handle
the error condition for fgets(), which may leave the buffer in an
undefined state in error.  Define the state of the buffer on error
by initializing the first character to a NULL terminator, which
will trip the already existing error handling.

Second, handle errors from a call to hwloc that could cause
very large allocations (due to -1 being cast to a size_t).

Signed-off-by: Brian Barrett <bbarrett@amazon.com>