common/ompio and sharedfp/lockedfile: store full path
This fix is based on a bug report on the mailing list. If the user
opens a file in a directory, but executes a chdir later before closing the file,
ompio was unable to properly clean up the lock files generated by the lockedfile
component. After some investigation it was confirmed, that the same problem
arises if a user would perform the same sequence of operations using the MODE_DELETE_AFTER_CLOSE,
hence the fix also had to be applied to the general file handle, not just the sharedfp component.

Note, this fix should apply clearly to 5.0, but I am not entirely sure whether we'll have to generate
a separate patch for 4.1 and 4.0, since the common_ompio datastructure has changed and got an additional element.

Fixes issue #9924

Signed-off-by: Edgar Gabriel <edgar.gabriel1@outlook.com>

incorporate changes requested during the review

Signed-off-by: Edgar Gabriel <edgar.gabriel1@outlook.com>
(cherry picked from commit 6178e43a6bbdad703895401ad1428479d79d621e)