Define the object class for the isendrecv request once.
This patch works, but the code is ugly, each file using a static
structure and function which ends up being defined twice (for MPI and
PMPI API). There is no better solution at the MPI API level that would
not require adding additional files to move the structure declaration,
callback implementations and the class implementation into.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>