sessions:move opal_sma_base_f to ompi_framework
dependencies array.

If this is not done, then when using a non-ob1 PML, the stages of finalization
have the SMSC framework closed before the btl framework is closed or del_procs invoked
for the btl(s).

When using the --enable-mca_dso this results in a segfault as the code section for the SMSC
framework/components has been dlclosed.

Moving the opal_smc_base_framework from the opal_init_frameworks to the
ompi_framework_dependencies fixes this problem.

Related to #9097

Signed-off-by: Howard Pritchard <howardp@lanl.gov>
(cherry picked from commit ec21585ec1940f79d887beec5c48f53e9fc728f6)