Allow name query for all predefined NULL MPI objects.
This applies to MPI_DATATYPE_NULL, MPI_COMM_NULL and MPI_WIN_NULL.

MPI 4.0 states that "The predefined constant MPI_COMM_NULL is the
value used for invalid communicator handles." Hence, we might assume
that MPI_COMM_NULL is an invalid communicator, but that is only
partially true because in 4.12.4 it is clearly stated that the
translation function, MPI_*_C2F and MPI_*_F2C, should treat
MPI_COMM_NULL as a valiid communicator. So, the real role of these
MPI_*_NULL objects might be to point to valid MPI objects used as
bounds, but they represent valid MPI objects that cannot be used for
communications. In any case, as predefined handles they have an
associated name, and the user _must_ be able to get this name.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>