Fix a performance regression in datatype pack/unpack
In introducing native types to the pack/unpack methods in
30639166f60606454afb2d7b23e29f9523487c31, a set of divisions
were introduced that the compiler cannot optimize out. This
results in poor performance for small datatypes, where 4 divisions
are executed for each chunk. This patch restores pack
performance from 350MB/s to 1.5GB/s on a datatype with 4 doubles
in 2 cache lines.

The patch moves alignment checks and stride computation into macros
where we have access to the datatypes, which can be used by the
compiler to optimize the divisions/modulos.

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>