ompi/contrib: remove libompitrace
This small library was added for a specific purpose several years ago.
There is no longer a maintainer for this library, and it hasn't been
documented for years (and therefore there are likely very few -- if
any -- users).  We're therefore removing libompitrace from future
versions of Open MPI; it can be obtained from prior version of Open
MPI if it is desired.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>