osc/rdma: Fix priority setting bug
In a previous commit, I accidently set the priority to always
be 0.  Which is wrong, but had no practical impact, given
the priorities and availabilities of other OSC components.
However, it was wrong and this patch fixes the issue.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>
(cherry picked from commit cf253bc4293b76fdbf1bc180fec8e7edde2d269d)