Accept a NULL pointer to descriptor in sendi.
In the initial design _sendi would either send a immediate (aka very
short) message or, in case of resource exhaustion, return a descriptor
on the current BTL to host the message. It seems this logic was changed
sometimes recently, and now _sendi should not return a descriptor if the
immediate communication failed and the upper level does not want a
descriptor.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>