common/ucx: fix variable registration
The registration code in common/ucx was overly complex. It assumed
that re-registration is not desireable when it essentially does the
same as the find before the registration. The only thing that needs
protection is resetting the char * variables before they are
registered.

The primary purpose of this commit to fix two memory leaks caused
by the extra allocations to hold the strings. These allocations
are unnecessary.

Signed-off-by: Nathan Hjelm <hjelmn@google.com>
(cherry picked from commit 5ad772c1c60d95cd9e25433388df5929ae84ab1c)