docs: show version at top of first page
Having the Open MPI version at the top of the front index page is a
good, obvious visual indicator to the user to which version these docs
apply.

Usually, the version we display is derived from the contents of the
VERSION file.  However, when releasing Open MPI, we typically go back
and create an annotated git tag at a commit prior to the head of the
release branch.  This means that there is no git commit where the
VERSION file has an empty "greek" version.

As such, use the following scheme to determine what version string to
use for the RST macro "|ompi_ver|":

1. If there is an environment variable named READTHEDOCS_VERSION, use
   that (this env variable is documented here:
   https://docs.readthedocs.io/en/stable/builds.html#build-environment)
2. Otherwise, use the contents of the Open MPI VERSION file

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>