common/ompio: fix a division by zero problem in cost_calc
the cost_calc function is used to tune the number of aggregators used for collective I/O. A division by zero could occur at one location where two integeres are divided before being converted to a float. Reverting the sequence to explicitely convert the values first to floats before performing the division (as done at a number of other places in the same function) fixes the issue.

Fixes Issue #10115

Signed-off-by: Edgar Gabriel <edgar.gabriel1@outlook.com>