build: Swap CHECK_PACKAGE macros
Swap usage of OPAL_CHECK_PACKAGE for OAC_CHECK_PACKAGE across the code
base.  This is the bulk "easy" swap, which is mostly just removing
the now unnecessary preflight checks before calling CHECK_PACKAGE.
In a couple of situations, there were some indenting changes because
of the conditional check cleanups.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>