build: Add unique check to OPAL_SUMMARY_ADD
Rewrite OPAL_SUMMARY_ADD with the goal of better supporting the ability
to call SUMMARY_ADD with the same section and topic multiple times and
having SUMMARY_ADD do the unique check.

This patch also changes how strings are stored, so that there is
no longer a hacky substitution of the section and topic names to
replace spaces with underscores during storage and the opposite
during display, meaning that the strings were practically limited
to alphabetical characters and spaces.

This implementation does use slightly more memory and puts more
strings in the environment, but also has significantly fewer
corner-case behaviors.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>