build/dl: rework framework configure
Greatly simplify the dl framework configure by not trying to track
libs and by setting the flags/libs as <framework>_<component>_flag,
so that opal_mca pulls in all the wrapper flags automagically.
Because the individual components link against dl or ltdl and
components are always static, libopen-pal will depend on dl or ltdl
and libmpi will inherit those dependencies.  So we don't need all
the LIB tracking we were doing.

Update the package checks to use OAC_CHECK_PACKAGE instead of
OPAL_CHECK_PACKAGE, although it is unlikely either will ever
provide a pkg-config file.

Remove the backdoor disable-dl-dlopen flag, which is a duplicate of
--enable-mca-no-build=dl-dlopen.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>