osc/sm: Fix detection of non-contig info
Remove the info subscribe for both the alloc_shared_noncontig and
blocking_fence info keys.  This change was inspired by issue #10175,
which highlighted that we were not properly following the non-contig
info key (our behavior was standards compliant, but not particularly
helpful), because the info subscription was overwriting the provided
value.  In the investigation, it became clear that there is really
no advantage to subscribing to a key that can't be changed, so
drop the subscription code for the two keys that can't be changed
and fix a bug and remove code all at the same time.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>