build: move OPAL_CHECK_WITHDIR into java
There is only one file that uses the OPAL_CHECK_WITHDIR macro, the
java bindings check.  We want to keep the usage of OPAL_CHECK_WITHDIR
to a minimum, because it is essentially a duplicate of work done
elsewhere, so rename it and make it associated with the Java checks.
I'm sufficiently unsure about the Java checks that I do not feel
comfortable rewriting the macro to not use OPAL_CHECK_WITHDIR,
although I think someone with more Java knowledge could do so fairly
easily.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>