build: Fix message when pkg-config not found
Fix a bug in the autoconf macro that runs pkg-config so that it
properly evaluates the action-if-not-found when pkg-config itself is
not found.  Because of the way the tests were setup, this bug's only
impact was a missing answer to an AC_MSG_CHECKING, rather than
generating wrong build options.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>