wrapper: Fix mis-application of flags filter
Fix a bug introduced with 5d23e0, where we were filtering out standard
paths from preprocessor flags instead of compiler flags.  In practice,
this didn't change any behaviors because we wanted to filter
preprocessor flags (and already had, so this call was a no-op) and
we never put -Is in compiler flags.  But it's a mistake and Coverity
found it, so fix it.

Fixes CID 1503336.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>
(cherry picked from commit 810c3b1da4edff93a1bdc061986ade8f85354216)