configure.ac: group similar tests together
Add a missing AC_MSG about SPC, and group it together with the rest of
the MPI-layer options.  Also put the library naming test there, since
it emits a message about "libmpi_FOO".

Move the OSHMEM tests down below all of that so that the output in
that section solely has to do with OSHMEM tests (and not have the
results of some OMPI tests tacked on to the end of that section).

Finally, make some trivial whitespace fixes.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>