build: Run missing package dist-hook earlier
The hook that would fail "make dist" if packages weren't configured to
be included in the tarball ran at the end of "make dist".  This was
fine when the packages missing didn't break running make dist.  However,
Sphinx not being available means that "make dist" won't work because
of missing files, but the error was really not intuitive.

Move the dist-hook for missing packages from the top level to the
config directory Makefile, which has the practical outcome of running
the missing package check right after generation of the AUTHORS file,
well before any likely missing package make breakage, returning us to
clear error messages.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>