Populate some additional MPI_INFO_ENV keys with singleton launch.
Specifically, 'command', 'wdir' and 'argv'.

Pass down the argv and argc commands to the lower levels from MPI_Init().
In the singleton case, the PMIx calls to get these data will fail, so populate
them when they do if provided. Nothing to do for the MPI_Sessions model,
as it is not supported.

'wdir' can be populated with a good enough guess of getcwd().

Signed-off-by: Austen Lauria <awlauria@us.ibm.com>