docs: do not set HTML install tree to 0755 perms
Follow-on to 8ffe33f813: do not set the dirs in the installed HTML
files tree to have 0755 permissions because Automake doesn't do that
when it creates directories.  Instead, use "find ..." to find all
directories and files in the generated _built/html tree and use the
Automake-provided installer macros to create the target directories
and install the target files with desireable permissions.

Additionally, add a comment explaining why the install-data-hook rule
exists, and why we are doing what we are doing in it.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>
(cherry picked from commit 1867a7296c0e2670fef8fb11ffc8a2e79690099b)