odls/default: cap the max number of child FDs to close
On some versions of MacOS (e.g., 12.3.1), we have seen
sysconf(_SC_OPEN_MAX) -- and "ulimit -n" -- return very large numbers,
and sometime return -1 (which means "unlimited").  This can result in
an unreasonably large loop over closing all FDs (especially if -1 gets
interpreted as LONG_MAX).
https://github.com/open-mpi/ompi/issues/10358 has some links to others
who have seen this kind of behavior.

Add an MCA param (orte_odls_default_maxfd, defaulting to 1024) that
caps the max number of FDs to close in non-Linux environments.  Use an
MCA param because we're picking this max value fairly arbitrarily;
give users a way to change it someday, if needed.

Thanks to Scott Sayres for raising the issue.

This is not a cherry pick because this code (i.e., ORTE) no longer
exists on main.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>