Remove Trailing Whitespace
Removes trailing whitespace in one spot to allow IDE's to enable the auto-remove
trailing whitespace setting without adding a lot of extra diffs to
commits. When you make a 1 line change in a file and have 100 lines of
trailing whitespace changes in the diff, it is time consuming to review.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>