common/ucx: Fix case where variable goes out of scope
If a program makes the following call sequence:
```
MPI_T_init_thread
MPI_T_finalize
MPI_Init
```
Then a segv will occur in MPI_Init because the memory backing the
`*opal_common_ucx.tls` will have been freed by the `mca_base_var_finalize`
in `MPI_T_finalize`. But the 'opal_common_ucx.tls` will be valid.

So add a check for this scenario so that we properly initialize the
default string even in this case.

 * Add missing unlock

Signed-off-by: Joshua Hursey <jhursey@us.ibm.com>