Removed deadlock in find dynamic region
This is my solution to issue #10328. The problem with this code is that inside of `ompi_osc_rdma_refresh_dynamic_region` there is a `ompi_osc_rdma_lock_acquire_shared` to the same lock used in `ompi_osc_rdma_lock_acquire_exclusive` generation a deadlock. My solution was simply to remove the outermost lock. However, I don't know if this is the better solution, but, with these changes, my test codes ran as I expected.