btl/base: push operation->hdr to am_rdma_respond for queued operation
Currently, when calling am_rdma_respond() for a queued
operation, amd_rdma_retry_operation() pass NULL for the hdr argument.

The idea is that hdr is only used for allocating operation->descriptor.
A queued operation should already have a descriptor, therefore does
not need hdr.

This missed the possibility that the allocation of descriptor
in am_rdma_respond() can fail, which will lead to the operation
to be queued without a descriptor.

This patch make retry_operation() to pass operation->hdr to
am_rdma_repsond() to address the issue.

It also added an assertion in am_rdma_repsond() about hdr must
not be NULL before hdr is being used.

Signed-off-by: Wei Zhang <wzam@amazon.com>