C11 atomics: replace _Atomic with volatile
Using the _Atomic type modifier leads to sequentially consistent atomic
access on variables even in code paths where atomicity and ordering
is not needed (opal_thread_* for example). By replacing _Atomic
with volatile in the "public" interface of the C11 backend, we preserve
the ability to suppress certain compiler optimizations while allowing
for non-atomic accesses. All atomic accesses and ordering is done
through the use of the opal_atomic* API throughout the code base
so _Atomic is not needed.

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>