github-labeler: add label corresponding to the target branch
This is a small GitHub action and corresponding Python script which
automatically adds labels to pull requests on PR creation and edits (i.e., when
the base branch changes). It adds labels in the form of "Target: {base}",
corresponding to the base branch of the PR. (NOTE: it does NOT create labels, so
the maintainers are still in control of which labels exist).

If the PR is labeled incorrectly - either by a user choosing the wrong label, or
by the base branch changing - the action will remove the incorrect "Target:
{base}" label and add the correct one. If the PR is labeled correctly, no action
is taken.

Signed-off-by: Joe Downs <joe@dwns.dev>
(cherry picked from commit ff3ee15b7ffaafdc9f25c232ad8b1dcfcf61cf27)