ompi/dpm: make procs consistent before calling PMIx_Connect()
ompi_dpm_connect_accept() call PMIx_Connect() to establish
connection, with "procs" as an argument.

PMIx requires "procs" to be consistent accross clients.

When it is used to set up inter-communicator communication,
ompi_dpm_connect_accept() does not maintain the order of
proc in "procs". This is because the function is called
by both MPI_Comm_connect() and MPI_Comm_accept(), and
it always put processes in local communicator in "procs"
first, followed by processes in remote communicator.

However, for caller of MPI_Comm_connect() and MPI_Comm_accept(),
local communicator and remote communicator are different.

This patch fixed the issue by sorting "procs" before
it is used to call PMIx_Connect(), this ensures that "procs"
are consistent accross processes.

Signed-off-by: Wei Zhang <wzam@amazon.com>