fbtl/posix: fix length calculation for data sieving
This commit fixes the calculation of the buffer length that
needs to be read when using data sieving. The original code implicitely
assumed that the ub of an iov at index j+1 is larger than
the ub of the iov at index j. This is not necessarily the case
for read operations. Hence, the code needs to keep track of the max.
ub found.

Fixes issue #10546

Signed-off-by: Edgar Gabriel <edgar.gabriel1@outlook.com>
(cherry picked from commit 6891cee622071dbe83f080676698c9a40b0a435b)