sessions: fix arg checking for group_from_session_pset
The group handle to MPI_Group_from_session_pset is an OUT parameter as defined in the MPI 4.0 spec.
The spec is silent about what the value pointed to by the group handle argument can/should be.

Note the second example in section 11.4 of the specification fails unless this patch is applied.

Related to issue #10577

Signed-off-by: Howard Pritchard <hppritcha@gmail.com>
(cherry picked from commit 5408bab3403a4080480d78943e911e4548e6777b)