First commit of libevent 2.1.12
Replace libevent 2.0.22 with 2.1.12 to address some CVEs.  The CVEs in
question were in code that Open MPI did not use, and were therefore
harmless.  However, in order to not have to continually explain this
to end users, we'll just upgrade the embedded libevent to a version
that does not contain the CVEs at all.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>