Atomics: default to GCC builtin atomics and use C11 _Atomic as fall-back
The C11 atomics require annotating types with _Atomic so assignment
involves sequential consistency. The overhead can be significant and
defeats the purpose of OPAL_THREAD_*, which try to replace atomic
operations with non-atomic operations if multi-threading is not enabled.

This commit replaces C11 atomics with the GCC builtins (`__atomic_*`) and
uses C11 as a fallback. It also fixes a couple of issues in the builtin
atomic backend.

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>