OFI: call fi_getname twice
first to get the length of the buffer needed for the endpoint name,
then a second time with the properly sized buffer to receive the
endpoint name.

The FI_NAME_MAX enum was not supposed to have been exposed to libfabric consumers.
See https://github.com/ofiwg/libfabric/issues/7898.

Related to #10617

Signed-off-by: Howard Pritchard <howardp@lanl.gov>