Implement warning for use of MPI_Sizeof in Fortran applications in MPI 4.0
This is implemented by adding '!gcc$ attributes deprecated ::' directives
in the declaration of each variation of MPI_Sizeof specified in the
header files defining MPI_Sizeof as a Fortran interface.

These directives are generated only if the compiler used to build Open MPI
is gfortran 11.0 or later.

Earlier versions of gfortran do not recognize this directive and flag it
as a syntax error.

Fortran compilers other than gfortran are unlikely to recognize this directive.

Signed-off-by: David Wootton <dwootton@us.ibm.com>