build: Cache build mode for components
Create cache variables for component build mode (dso vs. static).
We have a use case for components knowing how other components are
built (specifically, components that depend on a common component),
and this is a generic way of exposing that.

At the same time, slowly start trying to be more safe about variable
usage in the macros, using the Autoconf polymorphic variable functions
where possible.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>