update to safety valve for MCA callbacks
I used to have a "use_safety_valve" that was meant to conditionally
turn on the extra "unregister" call at MCA .so unload time.  Reviewers
became conserned about whether this was module-safe and wanted locks,
but really the entire "use_safety_valve" bookkeeping is unnecessary.

The logic of a safety valve is to assume bad code (like the oshmem
mismatch) could be using the register/unregister system and thus to
always unconditionally call unregister at unload time, regardless of
whether a nice matched pair of register/deregister already happened or
if we're in the oshmem case where they failed to make such a match.

With that being the design it was just necessary to update the
opal_mem_hooks_unregister_release() function to be safe regardless of
what context it's being called from (eg, if the callback is already
unregistered, or if it was never registered in the first place).

Signed-off-by: Mark Allen <markalle@us.ibm.com>